Pull FreeTextFilterSearchBar into separate file, make it internal
